created basic diffuse material!
created diffuse material. Note: any algorithm that randomizes direction of rays on bounce will produce surfaces that look matte. If a ray bounces off of a material and keeps 100% of its color, then we say that the material is white. If a ray bounces off of a material and keeps 0% of its color, then we say that the material is black.

GOAL: move on to adding primitives soon