working on reading from JSON. Also implememted code profiling workflow
for code perf: added perf command in makefile to compile with callgrind, then run a python script gprof2dot.py to generate an svg that evaluates the program

for reading from JSON: using nlohmann's JSON lib. NOTE: will require modifying class so that all variables are publicly accessible.

Due to this, it makes sense to refactor the hpp file tosplit up the declaration and implementation of the functions, as is standard practice. Just to make code more readible

modified gitignore file to ignore .svg files