added default constructor to sphere, turned it into a struct for json conversion purposes
testing the nlohmann json lib and my conclusion  is that modifying the namespace for user-declared classes which depend on other user declared classes is too much of a hassle. If instead the class is default constructible, i can declare a from_json function within the global namespace of my code.